+++ CHANGELOG.md
@@ -40,6 +40,13 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 
 ### Added
 
+- Extraction subcommands of `code42 security-data`, `print/write-to/send-to` accepts argument `--saved-search` to
+   return saved search results.
+
+- `code42 security-data saved-search` commands:
+    - `list` prints out existing saved search's id and name
+    - `show` takes a search id
+
 - `code42 high-risk-employee bulk` supports `add-risk-tags` and `remove-risk-tags`.
     - `code42 high-risk-employee bulk generate-template <cmd>` options `add-risk-tags` and `remove-risk-tags`.
         - `add-risk-tags` that takes a csv file with username and space separated risk tags.